requirements.txt
django
pillow
easyocr
opencv-python
numpy
python-dotenv